+++ spec/controllers/v1/users_controller_spec.rb
@@ -30,7 +30,7 @@ RSpec.describe V1::UsersController, type: :controller do
   end
 
   describe "POST create" do
-    let(:valid_attributes) { attributes_for(:user, password: "something") }
+    let(:valid_attributes) { attributes_for(:user, :with_email, :with_password) }
 
     describe "with valid params" do
       it "creates a new User" do